Fritz/requirements.txt
numpy
gcn_kafka==0.3.3
pandas
astropy==7.0.1
astropy_healpix==1.1.0
scikit_learn==1.5.1
requests
xmltodict==0.14.2
ligo.skymap==2.2.2
healpy==1.18.0
matplotlib
joblib==1.2.0
dill==0.3.9
pre-commit
keras==3.4.1
tensorflow==2.16.1
